+++ src/XRP/serializer.ts
@@ -1,15 +1,27 @@
+import Utils from '../Common/utils'
+
 import { XRPDropsAmount } from './generated/org/xrpl/rpc/v1/amount_pb'
 import {
+  Memo,
   Payment,
   Transaction,
 } from './generated/org/xrpl/rpc/v1/transaction_pb'
-import Utils from '../Common/utils'
 
 interface PaymentJSON {
-  Amount: object | string
+  Amount: Record<string, unknown> | string
   Destination: string
   DestinationTag?: number
   TransactionType: string
+}
+
+interface MemoJSON {
+  Memo?: MemoDetailsJSON
+}
+
+interface MemoDetailsJSON {
+  MemoData?: Uint8Array
+  MemoType?: Uint8Array
+  MemoFormat?: Uint8Array
 }
 
 interface BaseTransactionJSON {
@@ -19,6 +31,7 @@ interface BaseTransactionJSON {
   Sequence: number
   SigningPubKey: string
   TxnSignature?: string
+  Memos?: MemoJSON[]
 }
 
 interface PaymentTransactionJSONAddition extends PaymentJSON {
@@ -33,30 +46,44 @@ export type TransactionJSON = BaseTransactionJSON | PaymentTransactionJSON
 /**
  * Provides functionality to serialize from protocol buffers to JSON objects.
  */
-class Serializer {
+const serializer = {
   /**
    * Convert a Transaction to a JSON representation.
    *
-   * @param {proto.Transaction} transaction A Transaction to convert.
-   * @param signature An optional hex encoded signature to include in the transaction.
-   * @returns {Object} The Transaction as JSON.
+   * @param transaction - A Transaction to convert.
+   * @param signature - An optional hex encoded signature to include in the transaction.
+   * @returns The Transaction as JSON.
    */
-  public static transactionToJSON(
+  // eslint-disable-next-line max-statements, max-lines-per-function -- No clear way to refactor this because gRPC is verbose.
+  transactionToJSON(
     transaction: Transaction,
     signature?: string,
   ): TransactionJSON | undefined {
     const object: TransactionJSON = {
       Account: '',
-      Sequence: 0,
-      SigningPubKey: '',
-      LastLedgerSequence: 0,
       Fee: '',
+      Sequence: 0,
+      LastLedgerSequence: 0,
+      SigningPubKey: '',
     }
 
-    const sequence = transaction.getSequence()?.getValue()
-    if (sequence) {
-      object.Sequence = sequence
+    const normalizedAccount = getNormalizedAccount(transaction)
+    if (!normalizedAccount) {
+      return undefined
     }
+    object.Account = normalizedAccount
+
+    // Convert XRP denominated fee field.
+    const txFee = transaction.getFee()
+    if (txFee === undefined) {
+      return undefined
+    }
+    object.Fee = this.xrpAmountToJSON(txFee)
+
+    // Set sequence numbers
+    object.Sequence = transaction.getSequence()?.getValue() ?? 0
+    object.LastLedgerSequence =
+      transaction.getLastLedgerSequence()?.getValue() ?? 0
 
     const signingPubKeyBytes = transaction
       .getSigningPublicKey()
@@ -65,73 +92,29 @@ class Serializer {
       object.SigningPubKey = Utils.toHex(signingPubKeyBytes)
     }
 
-    const lastLedgerSequence = transaction.getLastLedgerSequence()?.getValue()
-    if (lastLedgerSequence) {
-      object.LastLedgerSequence = lastLedgerSequence
+    const additionalTransactionData = getAdditionalTransactionData(transaction)
+    if (additionalTransactionData === undefined) {
+      return undefined
     }
-
-    // Convert account field, handling X-Addresses if needed.
-    const accountAddress = transaction.getAccount()
-    if (!accountAddress) {
-      return
-    }
-    const account = accountAddress.getValue()?.getAddress()
-    if (!account || !Utils.isValidAddress(account)) {
-      return
-    }
-
-    let normalizedAccount = account
-    if (Utils.isValidXAddress(account)) {
-      const decodedClassicAddress = Utils.decodeXAddress(account)
-      if (!decodedClassicAddress) {
-        return
-      }
-
-      // Accounts cannot have a tag.
-      if (decodedClassicAddress.tag !== undefined) {
-        return
-      }
-
-      normalizedAccount = decodedClassicAddress.address
-    }
-    object.Account = normalizedAccount
-
-    // Convert XRP denominated fee field.
-    const txFee = transaction.getFee()
-    if (txFee === undefined) {
-      return
-    }
-    object.Fee = this.xrpAmountToJSON(txFee)
-
-    // Convert additional transaction data.
-    const transactionDataCase = transaction.getTransactionDataCase()
-    switch (transactionDataCase) {
-      case Transaction.TransactionDataCase.PAYMENT: {
-        const payment = transaction.getPayment()
-        if (payment === undefined) {
-          return
-        }
-        Object.assign(object, this.paymentToJSON(payment))
-        break
-      }
-      default:
-        throw new Error('Unexpected transactionDataCase')
-    }
+    Object.assign(object, additionalTransactionData)
 
     if (signature) {
       object.TxnSignature = signature
     }
 
+    Object.assign(object, this.memosToJSON(transaction.getMemosList()))
+
     return object
-  }
+  },
 
   /**
    * Convert a Payment to a JSON representation.
    *
-   * @param {proto.Payment} payment The Payment to convert.
-   * @returns {Object} The Payment as JSON.
+   * @param payment - The Payment to convert.
+   * @returns The Payment as JSON.
    */
-  private static paymentToJSON(payment: Payment): object | undefined {
+  // eslint-disable-next-line max-statements -- No clear way to make this more succinct because gRPC is verbose
+  paymentToJSON(payment: Payment): PaymentJSON | undefined {
     const json: PaymentJSON = {
       Amount: {},
       Destination: '',
@@ -139,48 +122,125 @@ class Serializer {
     }
 
     // If an x-address was able to be decoded, add the components to the json.
-    const destinationAddress = payment.getDestination()
-    if (!destinationAddress) {
-      return
-    }
-
-    const destination = destinationAddress.getValue()?.getAddress()
+    const destination = payment.getDestination()?.getValue()?.getAddress()
     if (!destination) {
-      return
+      return undefined
     }
 
     const decodedXAddress = Utils.decodeXAddress(destination)
-    if (!decodedXAddress) {
-      json.Destination = destination
-      delete json.DestinationTag
-    } else {
-      json.Destination = decodedXAddress.address
-      if (decodedXAddress.tag !== undefined) {
-        json.DestinationTag = decodedXAddress.tag
-      }
+    json.Destination = decodedXAddress?.address ?? destination
+    if (decodedXAddress?.tag !== undefined) {
+      json.DestinationTag = decodedXAddress.tag
     }
 
-    const amount = payment.getAmount()
-    if (!amount) {
-      return
-    }
-    const xrpAmount = amount.getValue()?.getXrpAmount()
+    const xrpAmount = payment.getAmount()?.getValue()?.getXrpAmount()
     if (!xrpAmount) {
-      return
+      return undefined
     }
     json.Amount = this.xrpAmountToJSON(xrpAmount)
+
     return json
-  }
+  },
 
   /**
    * Convert an XRPDropsAmount to a JSON representation.
    *
-   * @param xrpDropsAmount The XRPAmount to convert.
+   * @param xrpDropsAmount - The XRPAmount to convert.
    * @returns The XRPAmount as JSON.
    */
-  private static xrpAmountToJSON(xrpDropsAmount: XRPDropsAmount): string {
+  xrpAmountToJSON(xrpDropsAmount: XRPDropsAmount): string {
     return `${xrpDropsAmount.getDrops()}`
-  }
+  },
+
+  /**
+   * Convert an array of Memo objects to a JSON representation keyed by
+   * a field called 'Memos' iff the array is not empty and it contains
+   * non-empty objects.
+   *
+   * @param memos - The Memos to convert.
+   *
+   * @returns An array of the Memos in JSON format, or undefined.
+   */
+  memosToJSON(memos: Memo[]): { Memos: MemoJSON[] } | undefined {
+    if (!memos.length) {
+      return undefined
+    }
+
+    const convertedMemos = memos.map((memo) => this.memoToJSON(memo))
+    return { Memos: convertedMemos }
+  },
+
+  /**
+   * Convert a Memo to a JSON representation.
+   *
+   * @param memo - The Memo to convert.
+   * @returns The Memo as JSON.
+   */
+  memoToJSON(memo: Memo): MemoJSON {
+    const jsonMemo: MemoDetailsJSON = {
+      MemoData: memo.getMemoData()?.getValue_asU8(),
+      MemoFormat: memo.getMemoFormat()?.getValue_asU8(),
+      MemoType: memo.getMemoType()?.getValue_asU8(),
+    }
+
+    return {
+      Memo: jsonMemo,
+    }
+  },
 }
 
-export default Serializer
+export default serializer
+
+/**
+ * Converts the transaction's account field, handling X-Addresses if needed.
+ *
+ * @param transaction - The transaction to scrape the account from.
+ *
+ * @returns A XRP address or undefined.
+ */
+function getNormalizedAccount(transaction: Transaction): string | undefined {
+  const account = transaction.getAccount()?.getValue()?.getAddress()
+  if (!account || !Utils.isValidAddress(account)) {
+    return undefined
+  }
+
+  if (Utils.isValidClassicAddress(account)) {
+    return account
+  }
+
+  // We already checked that we're a valid address, and if we were a classic address,
+  // so we are definitely an X-Address here.
+  const decodedClassicAddress = Utils.decodeXAddress(account)
+  if (!decodedClassicAddress || decodedClassicAddress.tag !== undefined) {
+    // Accounts cannot have a tag.
+    return undefined
+  }
+
+  return decodedClassicAddress.address
+}
+
+/**
+ * Given a Transaction, get the JSON representation of data specific to that transaction type.
+ *
+ * @param transaction - A transaction to check for additional transaction-type specific data.
+ * @returns A JSON representation of the transaction-type specific data, or undefined.
+ */
+function getAdditionalTransactionData(
+  transaction: Transaction,
+): PaymentJSON | undefined {
+  const transactionDataCase = transaction.getTransactionDataCase()
+
+  switch (transactionDataCase) {
+    case Transaction.TransactionDataCase.PAYMENT: {
+      const payment = transaction.getPayment()
+      if (payment === undefined) {
+        return undefined
+      }
+
+      return serializer.paymentToJSON(payment)
+    }
+
+    default:
+      throw new Error('Unexpected transactionDataCase')
+  }
+}