+++ test/Common/utils.test.ts
@@ -1,4 +1,5 @@
 import { assert } from 'chai'
+
 import Utils from '../../src/Common/utils'
 import 'mocha'
 
@@ -8,10 +9,10 @@ describe('utils', function (): void {
     const address = 'rU6K7V3Po4snVhBBaU29sesqs2qTQJWDw1'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed valid.
-    assert.isTrue(validAddress)
+    assert.isTrue(isValidAddress)
   })
 
   it('isValidAddress() - Valid X-Address', function (): void {
@@ -19,10 +20,10 @@ describe('utils', function (): void {
     const address = 'XVLhHMPHU98es4dbozjVtdWzVrDjtV18pX8yuPT7y4xaEHi'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed valid.
-    assert.isTrue(validAddress)
+    assert.isTrue(isValidAddress)
   })
 
   it('isValidAddress() - Wrong Alphabet', function (): void {
@@ -30,10 +31,10 @@ describe('utils', function (): void {
     const address = '1EAG1MwmzkG6gRZcYqcRMfC17eMt8TDTit'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed invalid.
-    assert.isFalse(validAddress)
+    assert.isFalse(isValidAddress)
   })
 
   it('isValidAddress() - Invalid Classic Address Checksum', function (): void {
@@ -41,10 +42,10 @@ describe('utils', function (): void {
     const address = 'rU6K7V3Po4sBBBBBaU29sesqs2qTQJWDw1'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed invalid.
-    assert.isFalse(validAddress)
+    assert.isFalse(isValidAddress)
   })
 
   it('isValidAddress() - Invalid X-Address Checksum', function (): void {
@@ -52,10 +53,10 @@ describe('utils', function (): void {
     const address = 'XVLhHMPHU98es4dbozjVtdWzVrDjtV18pX8yuPT7y4xaEHI'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed invalid.
-    assert.isFalse(validAddress)
+    assert.isFalse(isValidAddress)
   })
 
   it('isValidAddress() - Invalid Characters', function (): void {
@@ -63,10 +64,10 @@ describe('utils', function (): void {
     const address = 'rU6K7V3Po4sBBBBBaU@#$%qs2qTQJWDw1'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed invalid.
-    assert.isFalse(validAddress)
+    assert.isFalse(isValidAddress)
   })
 
   it('isValidAddress() - Too Long', function (): void {
@@ -75,10 +76,10 @@ describe('utils', function (): void {
       'rU6K7V3Po4snVhBBaU29sesqs2qTQJWDw1rU6K7V3Po4snVhBBaU29sesqs2qTQJWDw1'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed invalid.
-    assert.isFalse(validAddress)
+    assert.isFalse(isValidAddress)
   })
 
   it('isValidAddress() - Too Short', function (): void {
@@ -86,10 +87,10 @@ describe('utils', function (): void {
     const address = 'rU6K7V3Po4s2qTQJWDw1'
 
     // WHEN the address is validated.
-    const validAddress = Utils.isValidAddress(address)
+    const isValidAddress = Utils.isValidAddress(address)
 
     // THEN the address is deemed invalid.
-    assert.isFalse(validAddress)
+    assert.isFalse(isValidAddress)
   })
 
   it('encodeXAddress() - Mainnet Address and Tag', function (): void {
@@ -152,6 +153,7 @@ describe('utils', function (): void {
   it('decodeXAddress() - Valid Mainnet Address with Tag', function (): void {
     // GIVEN an x-address that encodes an address and a tag.
     const address = 'XVfC9CTCJh6GN2x8bnrw3LtdbqiVCUvtU3HnooQDgBnUpQT'
+    const tag = 12345
 
     // WHEN it is decoded to an classic address
     const classicAddress = Utils.decodeXAddress(address)
@@ -161,13 +163,14 @@ describe('utils', function (): void {
       classicAddress!.address,
       'rU6K7V3Po4snVhBBaU29sesqs2qTQJWDw1',
     )
-    assert.strictEqual(classicAddress!.tag, 12345)
+    assert.strictEqual(classicAddress!.tag, tag)
     assert.strictEqual(classicAddress!.test, false)
   })
 
   it('decodeXAddress() - Valid Testnet Address with Tag', function (): void {
     // GIVEN an x-address that encodes an address and a tag.
     const address = 'TVsBZmcewpEHgajPi1jApLeYnHPJw82v9JNYf7dkGmWphmh'
+    const tag = 12345
 
     // WHEN it is decoded to an classic address
     const classicAddress = Utils.decodeXAddress(address)
@@ -177,7 +180,7 @@ describe('utils', function (): void {
       classicAddress!.address,
       'rU6K7V3Po4snVhBBaU29sesqs2qTQJWDw1',
     )
-    assert.strictEqual(classicAddress!.tag, 12345)
+    assert.strictEqual(classicAddress!.tag, tag)
     assert.strictEqual(classicAddress!.test, true)
   })