+++ test/XRP/wallet.test.ts
@@ -1,10 +1,12 @@
 import { assert } from 'chai'
+
 import Wallet from '../../src/XRP/wallet'
 import 'mocha'
 
 /**
  * A mapping of input and expected outputs for BIP39 and BIP44.
- * @see https://iancoleman.io/bip39/#english
+ *
+ * @see {@link https://iancoleman.io/bip39/#english|BIP39 - Mnemonic Code Converter}.
  */
 const derivationPathTestCases = {
   index0: {
@@ -17,7 +19,7 @@ const derivationPathTestCases = {
       '0090802A50AA84EFB6CDB225F17C27616EA94048C179142FECF03F4712A07EA7A4',
     expectedMainNetAddress: 'XVMFQQBMhdouRqhPMuawgBMN1AVFTofPAdRsXG5RkPtUPNQ',
     expectedTestNetAddress: 'TVHLFWLKvbMv1LFzd6FA2Bf9MPpcy4mRto4VFAAxLuNpvdW',
-    messageHex: Buffer.from('test message', 'utf-8').toString('hex'),
+    messageHex: Buffer.from('test message', 'utf8').toString('hex'),
     expectedSignature:
       '3045022100E10177E86739A9C38B485B6AA04BF2B9AA00E79189A1132E7172B70F400ED1170220566BD64AA3F01DDE8D99DFFF0523D165E7DD2B9891ABDA1944E2F3A52CCCB83A',
   },
@@ -43,7 +45,7 @@ describe('wallet', function (): void {
     assert.exists(walletGenerationResult!.mnemonic)
     assert.equal(
       walletGenerationResult!.derivationPath,
-      Wallet.getDefaultDerivationPath(),
+      Wallet.defaultDerivationPath,
     )
   })
 
@@ -61,7 +63,7 @@ describe('wallet', function (): void {
     )
     assert.equal(
       walletGenerationResult!.derivationPath,
-      Wallet.getDefaultDerivationPath(),
+      Wallet.defaultDerivationPath,
     )
     assert.equal(
       walletGenerationResult!.wallet.getAddress(),
@@ -84,7 +86,7 @@ describe('wallet', function (): void {
     )
     assert.equal(
       walletGenerationResult!.derivationPath,
-      Wallet.getDefaultDerivationPath(),
+      Wallet.defaultDerivationPath,
     )
     assert.equal(
       walletGenerationResult!.wallet.getAddress(),
@@ -111,8 +113,8 @@ describe('wallet', function (): void {
     )!
 
     // THEN the wallet has the expected address and keys.
-    assert.equal(wallet.getPrivateKey(), testData.expectedPrivateKey)
-    assert.equal(wallet.getPublicKey(), testData.expectedPublicKey)
+    assert.equal(wallet.privateKey, testData.expectedPrivateKey)
+    assert.equal(wallet.publicKey, testData.expectedPublicKey)
     assert.equal(wallet.getAddress(), testData.expectedMainNetAddress)
   })
 
@@ -128,8 +130,8 @@ describe('wallet', function (): void {
     )!
 
     // THEN the wallet has the expected address and keys.
-    assert.equal(wallet.getPrivateKey(), testData.expectedPrivateKey)
-    assert.equal(wallet.getPublicKey(), testData.expectedPublicKey)
+    assert.equal(wallet.privateKey, testData.expectedPrivateKey)
+    assert.equal(wallet.publicKey, testData.expectedPublicKey)
     assert.equal(wallet.getAddress(), testData.expectedTestNetAddress)
   })
 
@@ -144,8 +146,8 @@ describe('wallet', function (): void {
     )!
 
     // THEN the wallet has the expected address and keys.
-    assert.equal(wallet.getPrivateKey(), testData.expectedPrivateKey)
-    assert.equal(wallet.getPublicKey(), testData.expectedPublicKey)
+    assert.equal(wallet.privateKey, testData.expectedPrivateKey)
+    assert.equal(wallet.publicKey, testData.expectedPublicKey)
     assert.equal(wallet.getAddress(), testData.expectedAddress)
   })
 
@@ -157,8 +159,8 @@ describe('wallet', function (): void {
     const wallet = Wallet.generateWalletFromMnemonic(testData.mnemonic)!
 
     // THEN the wallet has the expected address and keys from the input mnemonic at the default derivation path.
-    assert.equal(wallet.getPrivateKey(), testData.expectedPrivateKey)
-    assert.equal(wallet.getPublicKey(), testData.expectedPublicKey)
+    assert.equal(wallet.privateKey, testData.expectedPrivateKey)
+    assert.equal(wallet.publicKey, testData.expectedPublicKey)
     assert.equal(wallet.getAddress(), testData.expectedMainNetAddress)
   })
 
@@ -191,10 +193,10 @@ describe('wallet', function (): void {
   it('walletFromSeed - TestNet', function (): void {
     // GIVEN a seed used to generate a wallet on TestNet
     const seed = 'snYP7oArxKepd3GPDcrjMsJYiJeJB'
-    const test = true
+    const isTest = true
 
     // WHEN a wallet is generated from the seed.
-    const wallet = Wallet.generateWalletFromSeed(seed, test)
+    const wallet = Wallet.generateWalletFromSeed(seed, isTest)
 
     // THEN the wallet has the expected address.
     assert.equal(